[LIB-62] Move rt worker to model/worker and init metatypes in data3Dtreemodel
