+++ MNE/disp3D/view3D.h
@@ -193,12 +193,6 @@ public:
 protected:
     //=========================================================================================================
     /**
-    * Init the meta types
-    */
-    void initMetatypes();
-
-    //=========================================================================================================
-    /**
     * Init the 3D view
     */
     void init();
